requirements.txt
google-api-python-client>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
